Working HashMap different class

+++ src/io/altar/jseproject/repository/ProductRepository.java
@@ -1,6 +1,7 @@
 package io.altar.jseproject.repository;
 
 import io.altar.jseproject.model.Product;
+import io.altar.jseproject.repository.EntityRepository;
 
 public class ProductRepository extends EntityRepository<Product> {
 	private static final ProductRepository INSTANCE = new ProductRepository();
@@ -11,7 +12,7 @@ public class ProductRepository extends EntityRepository<Product> {
 		return INSTANCE;
 	}
 
-	public static void alterElement(Integer id, Integer[] shelf, Integer discount, Integer tax, Double price) {
+	public static void alterElement(Integer id, Integer[] shelf, Double discount, Integer tax, Double price) {
 		((Product)ProductRepository.getInstance().get(id)).setShelf_idLoc(shelf);
 		((Product)ProductRepository.getInstance().get(id)).setDiscount(discount);
 		((Product)ProductRepository.getInstance().get(id)).setIva(tax);
